# docs/openapi.yaml
openapi: 3.0.3
info:
  title: URL Shortener API
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
...
  /api/redirect/{code}:
    get:
      summary: Redirect and record click
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect
        "404":
          description: Not found
...